Introduce parentheses
Major code changes:
- Add LPAREN and RPAREN tokens to lexer.py
- Add new grammar to parse parenthese in parser.py

Minor changes:
- Remove special $end token message because now I have no idea what it actually means
- Update README.md

So it turns out there is still an issue with $end. If there is ONE (and only one) newline at the end of test.txt, no errors are thrown. However, if there is a newline anywhere else or an extra one at the end, the $end error is thrown again.

On the plus side, parentheses are now a thing. I've managed to implement everything I've done so far in the vanilla-calc branch in less than a day, using RPly. Yeesh.